+++ client/src/components/Register/Login.css
@@ -1,0 +1,176 @@
+.login_section{
+    display: flex;
+    align-items:flex-start;
+    justify-content: space-around;
+    width: 40%;
+    margin-top: 3rem;
+}
+
+.login_container {
+    width: 80%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.login_form{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.login_h2, .login-form_label, .login-label_check, .login_link_terms, .link_recover_password, .login_with_google h3{
+    color: rgb(255, 255, 255);
+    text-shadow: 2px 2px 2px rgba(161, 161, 161, 0.6);
+}
+
+.login_h2{
+    font-size: 1.5rem;
+}
+
+.login_info_wraper{
+    display: flex;
+    flex-direction: column;
+    width: 80%;
+}
+
+.login_section_input{
+    background-color: rgb(255, 255, 255);
+    color: rgb(0, 0, 0);
+    font-size: 0.7rem;
+    width: 100%;
+    margin-bottom: 0.6rem;
+    height: 1.5rem;
+}
+
+.login_section_input::placeholder{
+    color: rgba(61, 60, 60, 0.85);
+    font-size: 0.7rem;
+    padding-left: 0.2rem;
+}
+
+.login-form_label {
+    font-size: 1rem; 
+    font-weight: 600;
+    width: 100%;
+    margin-bottom: 0.2rem;
+}
+
+.login_info_wraper_check{
+    display: flex;
+    height: 1.6rem;
+    width: 50%;
+    margin: 0.5rem;
+    justify-content: center;
+    justify-items: center;
+    justify-self: center;
+    align-items: center;
+}
+
+.login-label_check{
+    width: 85%;
+    height: 1.2rem;
+}
+
+.login_link_terms{
+    text-decoration: none;
+    font-size: 1rem;
+    font-size: 600;
+    transition: all 0.2s;
+    
+}
+
+.login_link_terms:hover{
+    color: rgb(239, 162, 60);
+}
+
+.login-input_check{
+    width: 10%;
+    height: 1.2rem;
+    margin: 0 0 0 0.5rem;
+ }
+
+.login_btn_div{
+    display: flex;
+    width: 50%;
+    background-color: transparent;
+    border: 1px solid white;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 3rem;
+    transition: all 0.2s;
+    border: none;
+}
+
+.login_btn_submit{
+    background-color: rgb(239, 162, 60);
+    width: 100%;
+    margin: 0.6rem 0;
+    padding: 0.4rem;
+    font-size: 1.2rem;
+    font-weight: 600;
+    border: none;
+}
+
+.login_btn_submit:hover {
+    background-color: #704782;
+    cursor: pointer;
+  }
+
+  .login_btn_text{
+    color: white;
+    font-size: 1.5rem;
+    margin-left: 0.5rem;
+  }
+
+  .link_recover_password{
+    margin-top: 0.5rem;
+    font: 0.7rem;
+    text-decoration: none;
+    transition: all 0.2s;
+  }
+
+  .link_recover_password:hover{
+    color: rgb(239, 162, 60);
+  }
+
+  .login_with_google{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+
+  .login_btn_google{
+    display: flex;
+    background-color: transparent;
+    border: 1px solid white;
+    place-items: center;
+    height: 3rem;
+    transition: 0.2s all;
+  }
+
+  .login_btn_google:hover{
+    background-color: #704782;
+    cursor: pointer;
+  }
+
+  .login_with_google h3{
+    font-size: 1.8rem;
+    font-weight: 600;
+  }
+
+  .login_with_google_logo{
+    width: 2.5rem;
+    height: auto;
+    transition: all 0.2s;
+  }
+
+  .login_with_google_logo:hover{
+    width: 2.8rem;
+    height: auto;
+  }
