+++ client/src/components/Register/Register.css
@@ -1,21 +1,12 @@
-.register_section{
+.register_page_container{
     display: flex;
-    align-items:flex-start;
-    justify-content: space-around;
     background-color: black;
     background-image: url('./bgRegister1.png');
     background-size: cover;
     width: 100%;
     height: 90.5vh;
-
-}
-
-.login_container, .register_container{
-    width: 45%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
+    align-items: self-start;
+    justify-content: space-between;
 }
 
 .sections_separation_div{
@@ -25,138 +16,3 @@
     background-color: rgb(239, 162, 60);
     border-radius: 0.5rem;
 }
-
-
-.register_form, .login_form{
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.register_h2, .register-form_label, .register-label_check, .register_link_terms, .link_recover_password, .login_with_google h3{
-    color: rgb(255, 255, 255);
-    text-shadow: 2px 2px 2px rgba(161, 161, 161, 0.6);
-}
-
-.register_h2{
-    font-size: 1.5rem;
-}
-
-.register_info_wraper{
-    display: flex;
-    flex-direction: column;
-    width: 50%;
-}
-
-.register_section_input{
-    background-color: rgb(255, 255, 255);
-    color: rgb(0, 0, 0);
-    font-size: 0.7rem;
-    width: 100%;
-    border-radius: 0.5rem;
-    margin-bottom: 0.6rem;
-}
-
-.register_section_input::placeholder{
-    color: rgba(61, 60, 60, 0.85);
-    font-size: 0.7rem;
-    padding-left: 0.2rem;
-}
-
-.register-form_label {
-    font-size: 1rem; 
-    font-weight: 600;
-    width: 100%;
-    margin-bottom: 0.2rem;
-}
-
-.register_info_wraper_check{
-    display: flex;
-    height: 1.6rem;
-    width: 50%;
-    margin: 0.5rem;
-    justify-content: center;
-    justify-items: center;
-    justify-self: center;
-    align-items: center;
-}
-
-.register-label_check{
-    width: 85%;
-    height: 1.2rem;
-}
-
-.register_link_terms{
-    text-decoration: none;
-    font-size: 1rem;
-    font-size: 600;
-    transition: all 0.2s;
-    
-}
-
-.register_link_terms:hover{
-    color: rgb(239, 162, 60);
-}
-
-.register-input_check{
-    width: 10%;
-    height: 1.2rem;
-    margin: 0 0 0 0.5rem;
- }
-
-.register_btn_div{
-    display: flex;
-    width: 50%;
-}
-
-.register_btn_submit{
-    background-color: rgb(239, 162, 60);
-    border-radius: 0.5rem;
-    width: 100%;
-    margin: 0.6rem 0;
-    padding: 0.4rem;
-    font-size: 1.2rem;
-    font-weight: 600;
-    transition: all 0.2s;
-}
-
-.register_btn_submit:hover {
-    background-color: #704782;
-    cursor: pointer;
-  }
-
-  .link_recover_password{
-    margin-top: 0.5rem;
-    font: 0.7rem;
-    text-decoration: none;
-    transition: all 0.2s;
-  }
-
-  .link_recover_password:hover{
-    color: rgb(239, 162, 60);
-  }
-
-  .login_with_google{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .login_with_google h3{
-    font-size: 1.8rem;
-    font-weight: 600;
-  }
-
-  .login_with_google_logo{
-    width: 2.5rem;
-    height: auto;
-    transition: all 0.2s;
-  }
-
-  .login_with_google_logo:hover{
-    width: 2.8rem;
-    height: auto;
-  }
