+++ client/src/index.js
@@ -1,3 +1,4 @@
+import './config/firebase-config';
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
